proto_c: regenerate protobuf C files and include OTel Profiles support
The following patch re-generate the protobuf C files and it now includes the
new (under development) profiles support.

The following versions has been used to regenerate the files:

- opentelemetry-proto: a8f08fc49 (Sep 26, 2024)
- protobuf-c: 8c201f6e (Nov 25, 2023)
  |--> protobuf version:       libprotoc 3.21.12

Note that protobuf-c has been patched to support optional fields, well not
actually support but allow to generate metrics.proto which uses it. The
patch is here:

https://github.com/cosmo0920/protobuf-c/commit/37506b514d994c0ff46615491d12a81a1dc1cb95

Signed-off-by: Eduardo Silva <eduardo@calyptia.com>